% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqcomplete.R
\name{dq_complete_plot}
\alias{dq_complete_plot}
\title{Create interactive percent completeness heatmap}
\usage{
dq_complete_plot(data, by_var, title, miss = NULL, margin = NULL)
}
\arguments{
\item{data}{the dataset}

\item{by_var}{the variable to stratify by}

\item{title}{title for your plot}

\item{miss}{an atomic vector listing what values to classify as missing (besides NA which will count as missing by default)}

\item{margin}{a list with margin (l:left, r:right, b:bottom, t:top) and padding sizes specified in pixels}
}
\value{
a plotly heatmap object
}
\description{
Create interactive percent completeness heatmap
}
